refactor: place Controls in separate component
