+++ src/pages/Game/Game.tsx
@@ -2,70 +2,38 @@ import React, { useState } from "react";
 import { PanelVoteCards } from "@/components/VoteCard";
 import { PanelUsers } from "@/components/User";
 import styles from './Game.module.scss';
-import { PrimaryButton, SecondaryButton } from "@/components/Button";
 import { CommonArea } from "@/components/CommonArea";
+import { Controls } from "@/components/Controls";
+import { Issues } from "@/components/Issues";
 
-const NAMES_BTN = {
-  run: "run round",
-  restart: "restart round",
-  stop: "stop game"
-};
 
 export const Game = () => {
   const title = 'Name room';
 
-  const [isCardVisible, setIsCardIsVisible] = useState(false);
+  const [isCardOpened, setIsCardIsVisible] = useState(false);
   const [selectedCardValue, setSelectedCardValue] = useState('');
-
-  const handlerRun = () => {
-    console.log("handlerRun");
-  };
-
-  const handlerRestart = () => {
-    console.log("handlerRestart");
-  };
-
-  const handlerStop = () => {
-    console.log("handlerStop");
-  };
 
   return (
     <main className={styles.main}>
+      <PanelUsers />
       <article className={styles.wrapper}>
-        <h1>{title}</h1>
-        <section className={styles.groupBtn}>
-          <PrimaryButton
-            className={styles.singleBtn}
-            onClick={handlerRun}
-          >
-            {NAMES_BTN.run}
-          </PrimaryButton>
-          <PrimaryButton
-            className={styles.singleBtn}
-            onClick={handlerRestart}
-          >
-            {NAMES_BTN.restart}
-          </PrimaryButton>
-          <SecondaryButton
-            className={styles.singleBtn}
-            onClick={handlerStop}
-          >
-            {NAMES_BTN.stop}
-          </SecondaryButton>
+        <section>
+          <h1 className={styles.title}>{title}</h1>
+          <Controls />
         </section>
-          <CommonArea 
-            isCardVisible={isCardVisible}
-            selectedCardValue={selectedCardValue}
-            setIsCardIsVisible={setIsCardIsVisible}
-            setSelectedCardValue={setSelectedCardValue}
-          />
+        <CommonArea 
+          isCardOpened={isCardOpened}
+          selectedCardValue={selectedCardValue}
+          setIsCardIsVisible={setIsCardIsVisible}
+          setSelectedCardValue={setSelectedCardValue}
+        />
         <PanelVoteCards 
           setSelectedCardValue={setSelectedCardValue}
           selectedCardValue={selectedCardValue}
-          isCardVisible={isCardVisible}
+          isCardOpened={isCardOpened}
         />
       </article>
-      <PanelUsers />
+      <Issues />
     </main>
   );
 };
